# app/openapi/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Team Task Manager API",
    "description": "Comprehensive API for team task management with authentication, analytics, and file attachments",
    "version": "2.0.0",
    "contact": {
      "name": "Team Task Manager",
      "email": "support@teamtaskmanager.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "SessionAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "REVIEW", "DONE"] },
          "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"] },
          "dueDate": { "type": "string", "format": "date-time" },
          "completedAt": { "type": "string", "format": "date-time" },
          "position": { "type": "number" },
          "teamId": { "type": "string" },
          "projectId": { "type": "string" },
          "assignedToId": { "type": "string" },
          "createdById": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": { "type": "array", "items": { "type": "object" } }
        }
      }
    }
  },
  "paths": {
    "/tasks": {
      "get": {
        "summary": "List tasks",
        "description": "Retrieve tasks with filtering and pagination",
        "parameters": [
          {
            "name": "teamId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "REVIEW", "DONE"] }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"] }
          },
          {
            "name": "assignedToId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "minimum": 0, "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tasks": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Task" }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "offset": { "type": "integer" },
                        "hasMore": { "type": "boolean" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create task",
        "description": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "teamId"],
                "properties": {
                  "title": { "type": "string", "minLength": 1, "maxLength": 200 },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "REVIEW", "DONE"], "default": "TODO" },
                  "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"], "default": "MEDIUM" },
                  "dueDate": { "type": "string", "format": "date-time" },
                  "teamId": { "type": "string" },
                  "projectId": { "type": "string" },
                  "assignedToId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "minLength": 1, "maxLength": 200 },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["TODO", "IN_PROGRESS", "REVIEW", "DONE"] },
                  "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"] },
                  "dueDate": { "type": "string", "format": "date-time" },
                  "assignedToId": { "type": "string" },
                  "position": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "summary": "Search across entities",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "minLength": 1 }
          },
          {
            "name": "entityTypes",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Comma-separated list of entity types (TASK,PROJECT,COMMENT)"
          },
          {
            "name": "teamId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "entityType": { "type": "string", "enum": ["TASK", "PROJECT", "COMMENT"] },
                          "title": { "type": "string" },
                          "description": { "type": "string" },
                          "relevanceScore": { "type": "number" }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "offset": { "type": "integer" },
                        "hasMore": { "type": "boolean" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}